pyproject.toml
[project]
name = "document-generator"
version = "0.1.0"
description = "A code analysis engine for document generation."
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "GitPython",
    "tree-sitter",
    "tree-sitter-language-pack",
    "pyyaml",
    "pytest",
    "httpx",
    "openai-agents>=0.6.1",
    "pytest-asyncio>=1.3.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "sqlalchemy>=2.0.44",
    "email-validator>=2.3.0",
    "ruff>=0.14.5",
]

[tool.uv]
# This section can be used for uv-specific configurations,
# but 'python' version is typically set in 'project.requires-python'.
# If you have specific uv settings, they would go here.

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "ruff",
    "black",
]

[tool.black]
line-length = 120
target-version = ['py312']
include = '\.pyi?$'
